% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_to_db.R
\name{response_to_db}
\alias{response_to_db}
\title{write response to database}
\usage{
response_to_db(credentials = NULL, db_table = "shiny_survey", response)
}
\arguments{
\item{credentials}{to connect to db (preferably from config.yml)}

\item{db_table}{table name in postgres database}

\item{response}{data to write (data.frame)}
}
\value{

}
\description{
write response to database
}
